Add BaseController for the API
in this way we can separate the logic used only in API's.
